requirement.txt

# Core dependencies
openai
langchain
fastapi
uvicorn
pydantic
python-dotenv
requests

# Database dependencies
supabase
psycopg2-binary
sqlalchemy
alembic

# Additional utilities
python-multipart
python-jose[cryptography]
passlib[bcrypt]
httpx
aiofiles
